# perf(agg): add benchmark for foldable aggregator