--- a/src/stream/src/lib.rs
+++ b/src/stream/src/lib.rs
@@ -41,6 +41,7 @@
 #![feature(iter_order_by)]
 #![feature(exact_size_is_empty)]
 #![feature(return_position_impl_trait_in_trait)]
+#![feature(test)]
 
 #[macro_use]
 extern crate tracing;