I hereby agree to the terms of the [RisingWave Labs, Inc. Contributor License Agreement](https://gist.github.com/TennyZhuang/f00be7f16996ea48effb049aa7be4d66#file-rw_cla).

## What's changed and what's your intention?

as title

<!--

**This section will be used as the commit message. Please do not leave this empty!**

Please explain **IN DETAIL** what the changes are in this PR and why they are needed:

- Summarize your change (**mandatory**)
- How does this PR work? Need a brief introduction for the changed logic (optional)
- Describe clearly one logical change and avoid lazy messages (optional)
- Describe any limitations of the current code (optional)
- Refer to a related PR or issue link (optional)

-->

## Checklist For Contributors

- [ ] I have written necessary rustdoc comments
- [ ] I have added necessary unit tests and integration tests
- [ ] I have added fuzzing tests or opened an issue to track them. (Optional, recommended for new SQL features #7934).
- [ ] I have demonstrated that backward compatibility is not broken by breaking changes and created issues to track deprecated features to be removed in the future. (Please refer to the issue)
- [ ] All checks passed in `./risedev check` (or alias, `./risedev c`)

## Checklist For Reviewers

- [ ] I have requested macro/micro-benchmarks as this PR can affect performance substantially, and the results are shown.
<!-- To manually trigger a benchmark, please check out [Notion](https://www.notion.so/risingwave-labs/Manually-trigger-nexmark-performance-dashboard-test-b784f1eae1cf48889b2645d020b6b7d3). -->

## Documentation

- [x] My PR **DOES NOT** contain user-facing changes.

benchmark
```
test bench_foldable_count_agg_vis_rate_0_05 ... bench:     189,442 ns/iter (+/- 1,419)
test bench_foldable_count_agg_vis_rate_0_25 ... bench:     209,099 ns/iter (+/- 1,080)
test bench_foldable_count_agg_vis_rate_0_5  ... bench:     236,969 ns/iter (+/- 614)
test bench_foldable_count_agg_vis_rate_0_75 ... bench:     271,424 ns/iter (+/- 2,308)
test bench_foldable_count_agg_without_vis   ... bench:     129,485 ns/iter (+/- 477)
test bench_foldable_max_agg_vis_rate_0_05   ... bench:     219,781 ns/iter (+/- 2,141)
test bench_foldable_max_agg_vis_rate_0_25   ... bench:     232,116 ns/iter (+/- 4,050)
test bench_foldable_max_agg_vis_rate_0_5    ... bench:     267,885 ns/iter (+/- 3,006)
test bench_foldable_max_agg_vis_rate_0_75   ... bench:     288,743 ns/iter (+/- 693)
test bench_foldable_max_agg_without_vis     ... bench:     257,916 ns/iter (+/- 448)
test bench_foldable_min_agg_vis_rate_0_05   ... bench:     220,225 ns/iter (+/- 2,015)
test bench_foldable_min_agg_vis_rate_0_25   ... bench:     235,350 ns/iter (+/- 1,213)
test bench_foldable_min_agg_vis_rate_0_5    ... bench:     261,232 ns/iter (+/- 1,756)
test bench_foldable_min_agg_vis_rate_0_75   ... bench:     289,783 ns/iter (+/- 967)
test bench_foldable_min_agg_without_vis     ... bench:     229,307 ns/iter (+/- 356)
test bench_foldable_sum_agg_vis_rate_0_05   ... bench:     235,178 ns/iter (+/- 1,452)
test bench_foldable_sum_agg_vis_rate_0_25   ... bench:     301,981 ns/iter (+/- 972)
test bench_foldable_sum_agg_vis_rate_0_5    ... bench:     405,590 ns/iter (+/- 3,253)
test bench_foldable_sum_agg_vis_rate_0_75   ... bench:     538,044 ns/iter (+/- 5,959)
test bench_foldable_sum_agg_without_vis     ... bench:     609,232 ns/iter (+/- 6,964)
```